Add some tests for handlebars translation. #181
