+++ common/test/common/templates/HandlebarsTemplateTest.java
@@ -5,19 +5,23 @@ import com.github.jknack.handlebars.io.TemplateLoader;
 import common.pages.*;
 import org.junit.Test;
 
+import java.util.List;
+import java.util.Locale;
+
 import static java.util.Arrays.asList;
 import static java.util.Collections.singletonList;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-public class HandlebarsTemplateServiceTest {
+public class HandlebarsTemplateTest {
     private static final TemplateLoader DEFAULT_LOADER = new ClassPathTemplateLoader("/templates");
     private static final TemplateLoader OVERRIDE_LOADER = new ClassPathTemplateLoader("/templates/override");
     private static final TemplateLoader WRONG_LOADER = new ClassPathTemplateLoader("/templates/wrong");
+    private static final List<Locale> LOCALES = singletonList(Locale.ENGLISH);
 
     @Test
     public void rendersTemplateWithPartial() throws Exception {
-        final String html = handlebars().render("template", pageDataWithTitleAndMessage());
+        final String html = handlebars().render("template", pageDataWithTitleAndMessage(), LOCALES);
         assertThat(html).contains("<title>foo</title>")
                 .contains("<h1>bar</h1>")
                 .contains("<h2></h2>")
@@ -27,7 +31,7 @@ public class HandlebarsTemplateServiceTest {
 
     @Test
     public void rendersOverriddenTemplateUsingOverriddenAndDefaultPartials() throws Exception {
-        final String html = handlebarsWithOverride().render("template", pageDataWithTitleAndMessage());
+        final String html = handlebarsWithOverride().render("template", pageDataWithTitleAndMessage(), LOCALES);
         assertThat(html).contains("overridden template")
                 .contains("overridden partial")
                 .contains("another default partial");
@@ -35,7 +39,7 @@ public class HandlebarsTemplateServiceTest {
 
     @Test
     public void rendersDefaultTemplateUsingOverriddenAndDefaultPartials() throws Exception {
-        final String html = handlebarsWithOverride().render("anotherTemplate", pageDataWithTitleAndMessage());
+        final String html = handlebarsWithOverride().render("anotherTemplate", pageDataWithTitleAndMessage(), LOCALES);
         assertThat(html).contains("default template")
                 .contains("overridden partial")
                 .contains("another default partial");
@@ -43,13 +47,13 @@ public class HandlebarsTemplateServiceTest {
 
     @Test
     public void throwsExceptionWhenTemplateNotFound() throws Exception {
-        assertThatThrownBy(() -> handlebars().render("unknown", pageDataWithTitleAndMessage()))
+        assertThatThrownBy(() -> handlebars().render("unknown", pageDataWithTitleAndMessage(), LOCALES))
                 .isInstanceOf(TemplateNotFoundException.class);
     }
 
     @Test
     public void usesFallbackContextWhenMissingData() throws Exception {
-        final String html = handlebarsWithFallbackContext(DEFAULT_LOADER).render("template", pageDataWithTitleAndMessage());
+        final String html = handlebarsWithFallbackContext(DEFAULT_LOADER).render("template", pageDataWithTitleAndMessage(), LOCALES);
         assertThat(html).contains("<title>foo</title>")
                 .contains("<h1>bar</h1>")
                 .contains("<h2>fallback unknown</h2>")
@@ -59,12 +63,24 @@ public class HandlebarsTemplateServiceTest {
 
     @Test
     public void failsSilentlyWhenFallbackContextNotFound() throws Exception {
-        final String html = handlebarsWithFallbackContext(WRONG_LOADER).render("template", pageDataWithTitleAndMessage());
+        final String html = handlebarsWithFallbackContext(WRONG_LOADER).render("template", pageDataWithTitleAndMessage(), LOCALES);
         assertThat(html).contains("<title>foo</title>")
                 .contains("<h1>bar</h1>")
                 .contains("<h2></h2>")
                 .contains("<p>default partial</p>")
                 .contains("<ul></ul>");
+    }
+
+    @Test
+    public void simpleTranslation() throws Exception {
+        final String html = handlebarsWithFallbackContext(WRONG_LOADER).render("translations/simple", pageDataWithTitleAndMessage(), LOCALES);
+        assertThat(html).contains("Sales Tax");
+    }
+
+    @Test
+    public void simpleTranslationWithBundle() throws Exception {
+        final String html = handlebarsWithFallbackContext(WRONG_LOADER).render("translations/simpleBundle", pageDataWithTitleAndMessage(), LOCALES);
+        assertThat(html).contains("Secure Checkout - Confirmation");
     }
 
     private TemplateService handlebars() {
