+++ common/test/common/templates/TemplateServiceTest.java
@@ -4,6 +4,9 @@ import common.pages.PageData;
 import org.junit.Test;
 import play.twirl.api.Html;
 
+import java.util.Locale;
+
+import static java.util.Collections.singletonList;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class TemplateServiceTest {
@@ -11,8 +14,8 @@ public class TemplateServiceTest {
     @Test
     public void fillsToHtml() throws Exception {
         final String expectedHtml = someHtml();
-        final TemplateService templateService = (t, p) -> expectedHtml;
-        final Html html = templateService.renderToHtml(someTemplateName(), somePageData());
+        final TemplateService templateService = (t, p, l) -> expectedHtml;
+        final Html html = templateService.renderToHtml(someTemplateName(), somePageData(), singletonList(Locale.ENGLISH));
         assertThat(html.body()).isEqualTo(expectedHtml);
     }
