+++ common/test/resources/templates/translations/simple.hbs
@@ -1,0 +1,1 @@
+{{i18n "common.salesTax"}}
