+++ common/test/resources/templates/translations/simpleBundle.hbs
@@ -1,0 +1,1 @@
+{{i18n "checkout:confirmationPage.title"}}
