+++ conf/application.conf
@@ -26,7 +26,10 @@ handlebars.templateLoaders=[
 handlebars.fallbackContexts=[
   {"type":"classpath", "path":"/META-INF/resources/webjars/templates"}
 ]
-
+handlebars.i18n {
+  langs=${play.i18n.langs}
+  bundles=["translations", "home", "catalog", "checkout"]
+}
 
 # common
 common.saleCategoryExternalId = "6"
